requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
Cython==0.29.28
Flask==2.0.2
Flask-Mail==0.9.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.4
opencv-python==4.5.5.64
packaging==21.3
pandas==1.3.4
Pillow==9.1.0
pyparsing==3.0.8
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
tesseract==0.1.3
threadpoolctl==3.0.0
urllib3==1.26.9
Werkzeug==2.0.2
wincertstore==0.2
